+++ source/extensions/omni.isaac.orbit_envs/omni/isaac/orbit_envs/classic/cartpole/cartpole_env.py
@@ -25,12 +25,12 @@ class CartpoleEnv(IsaacEnv):
         https://github.com/openai/gym/blob/master/gym/envs/classic_control/cartpole.py
     """
 
-    def __init__(self, cfg: dict, headless: bool = False):
+    def __init__(self, cfg: dict, **kwargs):
         """Initializes the environment.
 
         Args:
             cfg (dict): The configuration dictionary.
-            headless (bool, optional): Whether to enable rendering or not. Defaults to False.
+            kwargs (dict): Additional keyword arguments. See IsaacEnv for more details.
         """
         # copy configuration
         self.cfg_dict = cfg.copy()
@@ -40,7 +40,7 @@ class CartpoleEnv(IsaacEnv):
         )
         isaac_cfg.sim.from_dict(self.cfg_dict["sim"])
         # initialize the base class to setup the scene.
-        super().__init__(isaac_cfg, headless=headless)
+        super().__init__(isaac_cfg, **kwargs)
 
         # define views over instances
         self.cartpoles = ArticulationView(prim_paths_expr=self.env_ns + "/.*/Cartpole", reset_xform_properties=False)