+++ source/extensions/omni.isaac.orbit_envs/omni/isaac/orbit_envs/manipulation/lift/lift_env.py
@@ -27,7 +27,7 @@ from .lift_cfg import LiftEnvCfg, RandomizationCfg
 class LiftEnv(IsaacEnv):
     """Environment for lifting an object off a table with a single-arm manipulator."""
 
-    def __init__(self, cfg: LiftEnvCfg = None, headless: bool = False):
+    def __init__(self, cfg: LiftEnvCfg = None, **kwargs):
         # copy configuration
         self.cfg = cfg
         # parse the configuration for controller configuration
@@ -38,7 +38,7 @@ class LiftEnv(IsaacEnv):
         self.object = RigidObject(cfg=self.cfg.object)
 
         # initialize the base class to setup the scene.
-        super().__init__(self.cfg, headless=headless)
+        super().__init__(self.cfg, **kwargs)
         # parse the configuration for information
         self._process_cfg()
         # initialize views for the cloned scenes